Create shared example file for dru issue collection builders, start adding tests, moved build_issues and build_request_issues into it
